fix: fixed typo and made modular debounce()
