+++ edit.html
@@ -6,7 +6,7 @@
         <title>글 수정하기</title>
         <link href="https://fonts.googleapis.com/css2?family=Noto+Sans+KR:wght@100..900&display=swap" rel="stylesheet">
         <link rel="stylesheet" href="css/edit.css">
-        <script src="js/edit.js" defer></script>
+        <script type="module" src="js/edit.js" defer></script>
     </head>
     <body>
         <main>
