+++ index.html
@@ -6,7 +6,8 @@
         <title>치영의 일기장</title>
         <link rel="stylesheet" href="https://fonts.googleapis.com/css2?family=Noto+Sans+KR:wght@100..900&display=swap" integrity="sha384-80Gx8/ISg4/YRCM3FBSI6ZISwyv1YS06ohM5kCGY9Hz149a6RiLWj8JtqzLrevPZ" crossorigin="anonymous">
         <link rel="stylesheet" href="css/index.css">
-        <script type="text/javascript" src="js/index.js" defer></script>
+        <script type="module" src="js/functool.js"></script>
+        <script type="module" src="js/index.js" defer></script>
     </head>
     <body class="noto-sans-kr">
         <div class="curtain"></div>
